pyproject.toml
[tool.poetry]
name = "leetcode_env"
version = "0.2"
description = "Leetcode evaluation environment for code generation agents"
authors = ["Beck LaBash <becklabash@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/GammaTauAI/leetcode-hard-gym"

[tool.poetry.dependencies]
python = "^3.9"  
beautifulsoup4 = "4.12.0"
gym = "0.26.2"
pydantic = "1.10.7"
python-dotenv = "1.0.0"
requests = "2.28.2"
python-leetcode = "1.2.1"
astunparse = "1.6.3"
pandas = "1.5.3"
pytest = "7.4.0"
ipykernel = "6.25.1"
langchain = { version = "0.0.268", optional = true }
openai = { version = "0.27.8", optional = true }
termcolor = { version = "2.3.0", optional = true }

[tool.poetry.extras]
llms = ["langchain", "openai", "termcolor"]

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"